Move events.py out of spawned/
