+++ nextline/plugin/plugins/registrars/trace_nos.py
@@ -1,5 +1,5 @@
+from nextline.events import OnEndTrace, OnStartTrace
 from nextline.plugin.spec import Context, hookimpl
-from nextline.spawned import OnEndTrace, OnStartTrace
 from nextline.types import TraceNo
