+++ nextline/spawned/__init__.py
@@ -26,18 +26,6 @@ __all__ = [
 import traceback
 
 from .commands import Command, PdbCommand
-from .events import (
-    Event,
-    OnEndCmdloop,
-    OnEndPrompt,
-    OnEndTrace,
-    OnEndTraceCall,
-    OnStartCmdloop,
-    OnStartPrompt,
-    OnStartTrace,
-    OnStartTraceCall,
-    OnWriteStdout,
-)
 from .runner import run
 from .types import QueueIn, QueueOut, RunArg, RunResult, Statement
